evaluation/requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
Pillow==11.0.0
tqdm==4.67.1
timm==0.6.12
diffusers==0.21.4
colorama==0.4.6
transformers==4.30.0
huggingface-hub==0.16.4
safetensors==0.6.2
filelock==3.19.1
regex==2025.10.22
requests==2.28.1 

